refoctor: move UniqueList class into common module
